Complete Project
Program is designed to provide recommendations for UAS Pilots on whether or not to fly their drone based on current weather conditions. The goal of the project was to learn the basics of using API's in Python and apply it to a real world problem. The program will locally save the user's ability to fly at night, in wet conditions, if the operator can fly beyond visual line of sight (BVLOS), and the maximum wind speed the drone can fly at. After the user enters the location of the flight, the program will then print a weather report and a recommendation if the flight is safe to fly based on the user's preferences and current weather.